Splitting: the tracer test so USM can own their own test
